attempt to fix intial load render bug
on initial load all teams displayed as if isPicked state is false...console logs show they are null at first, but switch to false. If anything is done to refresh the page they display correctly.

isPicked === null return was not updated as previously discussed, thought it might fix bug but it had no impact on result after updated.