continued drying of handleSubmit and reviewPicks functions
refactor to pass an entire array to reviewPicks and have it return an object with 3 keys that are arrays.

reviewPicks is now called for picks and statPicks is either array has length at the time of button click.

the remainder of logic is now continuing to happen inside the array.length if statement for both picks and statPicks. If picks are updated or new, the endpoint is then hit, and the results are used to populate isSubmitted. Containing all logic in these if statements makes it so that I no longer need to sort by type in reviewPicks (eliminates a lot of wet ass arrays that were floating around in last iteration - can reuse array for either a pick of statPick now within the reviewPicks function)

isSubmitted is the only state that needs to be set there, NOT picks or statPicks